bot/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.20.0
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
multidict==4.7.1
websockets==6.0
yarl==1.4.2
